vm-3-processor/correlator/requirements.txt
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
